backend/requirements.txt
fastapi~=0.115.6
uvicorn[standard]
psycopg2-binary~=2.9.10
python-dotenv~=1.0.1
bcrypt~=4.2.1
pydantic~=2.10.5
python-multipart~=0.0.20
stripe~=12.1.0